# docs/spec/api/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise IDP API
  description: >
    API for onboarding business processes, discovering available Azure AI
    Content Understanding analyzers, and triggering/reviewing document
    inference pipeline runs.


    This document is hand-written and authoritative: the FastAPI backend
    implements it, and the TypeScript client is generated from it.


    Authentication supports three caller modes: the `httpOnly`
    `cl_idp_session` browser cookie set by `POST /auth/login`, short-lived
    per-user bearer tokens returned by `POST /auth/token`, and a shared
    `SERVICE_API_TOKEN` bearer token for trusted workers and scripts.
  version: "0.2.0"
servers:
  - url: /api
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Start a browser session
      security: []
      description: >
        Validates an end user's email/password, returns their public
        profile, and sets the `httpOnly` `cl_idp_session` cookie used by
        first-party browser calls. Unlike `POST /auth/token`, the JWT is
        not returned in the response body.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed-in user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/token:
    post:
      tags: [auth]
      summary: Mint a user bearer token
      security: []
      description: >
        Validates an end user's email/password and returns a short-lived
        `Authorization: Bearer` token for scripts, CLI calls, and ad-hoc
        testing. No cookie is set; the token carries the caller's own
        role, so RBAC stays the same as their browser session.
      operationId: issueToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Short-lived bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [auth]
      summary: Get the current authenticated user
      description: >
        Resolves the caller from either the browser session cookie or a
        user bearer token and returns that user's public profile.
        Service-token callers are not associated with a user and receive
        `401`.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Send either a short-lived user token minted by `POST /auth/token`
        or the shared `SERVICE_API_TOKEN` as `Authorization: Bearer
        <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: cl_idp_session
      description: >
        `httpOnly` browser session cookie set by `POST /auth/login`.
  responses:
...
    Unauthorized:
      description: Authentication is required or the supplied credentials were invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [code, message]
      description: >
        Uniform error body returned with every 4xx and 5xx response, with
        the single exception of `/healthz`, which returns a `Health` body on
        503 so that probes get dependency detail rather than a generic
        error.
      properties:
        code:
          type: string
          description: >
            Stable machine-readable identifier, e.g. `process_not_found`,
            `duplicate_process_name`, `unsupported_file_type`,
            `file_too_large`, `too_many_pages`, `routing_analyzer_not_ready`,
            `unknown_field_path`, `job_not_failed`,
            `content_understanding_unavailable`.
          example: unsupported_file_type
        message:
          type: string
          description: Human-readable explanation, safe to show to a user.
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: >
            Optional structured context (e.g. the offending field path, the
            actual file size, the conflicting process ID).
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    TokenResponse:
      type: object
      required: [accessToken, expiresIn]
      description: >
        Response body for `POST /auth/token` — a short-lived bearer token.
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
          default: Bearer
        expiresIn:
          type: integer
          description: Seconds until `accessToken` expires.
    UserRole:
      type: string
      enum: ["IT Admin", Reviewer, "End User"]
      description: >
        Closed set of roles RBAC enforcement is based on.
    User:
      type: object
      required: [id, email, displayName, roleLabel, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        roleLabel:
          $ref: "#/components/schemas/UserRole"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...